app/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
pydantic-settings==2.7.0
redis==5.2.1
aio-pika==9.5.4
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
